divided into functions map
